remove weird padding
while we're missing the crown and logo, we should remove the padding.
I didn't want to directly edit _header.scss since this is a short-term
thing, so i'm putting this hacky block directly in _core.scss so we
know we need to delete it later when restoring the crown and logo.